requirements.txt
anaconda-client==1.7.2
anaconda-navigator==2.0.3
appdirs==1.4.4
argh==0.26.2
atomicwrites==1.4.0
backports.weakref==1.0.post1
bkcharts==0.2
black==19.10b0
boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
certifi==2020.12.5
clyent==1.2.2
comtypes==1.1.9
conda==4.10.3
conda-build==3.21.4
conda-verify==3.4.2
contextlib2==0.6.0.post1
cycler==0.10.0
cytoolz==0.11.0
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
future==0.18.2
gdb==0.0.1
h5py==2.10.0
HeapDict==1.0.1
htmlmin==0.1.12
ImageHash==4.2.1
ipdb==0.13.9
jdcal==1.4.1
json5==0.9.5
jupyter==1.0.0
llvmlite==0.36.0
locket==0.2.1
MarkupSafe==2.0.1
mccabe==0.6.1
menuinst==1.4.16
missingno==0.5.0
mistune==0.8.4
mkl-fft==1.3.0
mkl-service==2.3.0
mpmath==1.2.1
multimethod==1.6
multipledispatch==0.6.0
mypy-extensions==0.4.3
navigator-updater==0.2.1
olefile==0.46
pandas-profiling==3.1.0
parso==0.7.0
pathspec==0.7.0
patsy==0.5.1
pep8==1.7.1
phik==0.12.0
pkginfo==1.7.0
ply==3.11
pycosat==0.6.3
pycurl==7.43.0.6
pydantic==1.8.2
pyodbc===4.0.0-unsupported
pyreadline==2.1
pytest==6.2.3
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
QDarkStyle==2.8.1
QtPy==1.9.0
scikit-image==0.18.1
scikit-learn==1.0.1
simplegeneric==0.8.1
sip==4.19.13
statsmodels==0.12.2
tables==3.6.1
tangled-up-in-unicode==0.1.0
terminado==0.9.4
torch==1.9.1
torchaudio==0.9.1
torchvision==0.10.1
unicodecsv==0.14.1
visions==0.7.4
webencodings==0.5.1
widgetsnbextension==3.5.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.12.1
xlwings==0.23.0
xlwt==1.3.0
xmltodict==0.12.0
zict==2.0.0
zope.event==4.5.0
